# Make simplecov/rcov optional