We don't need to see coverage on every test run, and it takes a second or two every time.

We do want to see it on CI, so re-enable there.  Set COVERAGE=true in your test env vars if you would like to keep the current behaviour.

### Before

```
bin/rspec spec/models/candidate_spec.rb:24  17.05s user 6.01s system 97% cpu 23.701 total
```

### After
```
bin/rspec spec/models/candidate_spec.rb:24  9.75s user 4.84s system 95% cpu 15.233 total
```

## Guidance to review

Could someone please confirm that I've put the CI configuration in the right place, thank you.

## Link to Trello card

https://trello.com/c/0M84STpQ/695-speed-up-app-development