@elliotcm 2022-10-03
![](https://github.trello.services/images/mini-trello-icon.png) [Speed up app development](https://trello.com/c/0M84STpQ/695-speed-up-app-development)